create MetaculusClient class
